fixing types in hamming

+++ src/algorithms/hamming.c
@@ -9,7 +9,7 @@ int hamming_distance(uint8_t *vec1, uint8_t *vec2, size_t size) {
         return -1; // Error code for null pointers
     }
     
-    uint16_t distance = 0;
+    int distance = 0;
     for (size_t i = 0; i < size; i++) {
         distance += __builtin_popcount(vec1[i] ^ vec2[i]);
     }
@@ -23,12 +23,12 @@ double hamming_similarity(uint8_t *vec1, uint8_t *vec2, size_t size) {
         return -1.0; // Error code for null pointers
     }
     
-    uint16_t distance = hamming_distance(vec1, vec2, size);
+    int distance = hamming_distance(vec1, vec2, size);
     if (distance == -1) { // Check if hamming_distance returned an error
         return -1.0; // Propagate the error code
     }
 
-    uint16_t vector_dim = BINARY_EMBEDDING_DIM;
+    int vector_dim = BINARY_EMBEDDING_DIM;
     return (double)(vector_dim - distance) / (double)vector_dim;
 }
